requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
idna==3.3
numpy==1.22.3
pandas==1.4.2
pycparser==2.21
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
urllib3==1.26.9
